+++ src/app/loadout-builder/generated-sets/GeneratedSetItem.tsx
@@ -15,9 +15,13 @@ import { generateMixesFromPerks } from '../process';
 /**
  * Figure out which (if any) non-selected perks should be selected to get the chosen stat mix.
  */
-function identifyAltPerkChoicesForChosenStats(item: DimItem, chosenValues: number[]) {
+function identifyAltPerkChoicesForChosenStats(
+  item: DimItem,
+  chosenValues: number[],
+  assumeMasterwork: boolean
+) {
   let altPerks: DimPlug[] = [];
-  generateMixesFromPerks(item, (mix, plugs) => {
+  generateMixesFromPerks(item, assumeMasterwork, (mix, plugs) => {
     if (plugs && mix.every((val, index) => val === chosenValues[index])) {
       altPerks = plugs;
       return false;
@@ -35,6 +39,7 @@ export default function GeneratedSetItem({
   locked,
   statValues,
   itemOptions,
+  assumeMasterwork,
   addLockedItem,
   removeLockedItem
 }: {
@@ -42,13 +47,14 @@ export default function GeneratedSetItem({
   locked?: readonly LockedItemType[];
   statValues: number[];
   itemOptions: DimItem[];
+  assumeMasterwork: boolean;
   addLockedItem(lockedItem: LockedItemType): void;
   removeLockedItem(lockedItem: LockedItemType): void;
 }) {
-  const altPerks = useMemo(() => identifyAltPerkChoicesForChosenStats(item, statValues), [
-    item,
-    statValues
-  ]);
+  const altPerks = useMemo(
+    () => identifyAltPerkChoicesForChosenStats(item, statValues, assumeMasterwork),
+    [item, statValues, assumeMasterwork]
+  );
 
   const classesByHash = altPerks.reduce(
     (memo, perk) => ({ ...memo, [perk.plugItem.hash]: styles.altPerk }),