+++ src/app/loadout-builder/generated-sets/GeneratedSets.tsx
@@ -21,6 +21,7 @@ interface Props {
   statOrder: StatTypes[];
   defs: D2ManifestDefinitions;
   enabledStats: Set<StatTypes>;
+  assumeMasterwork: boolean;
   onLockedMapChanged(lockedMap: Props['lockedMap']): void;
 }
 
@@ -89,7 +90,8 @@ export default class GeneratedSets extends React.Component<Props, State> {
       isPhonePortrait,
       combos,
       combosWithoutCaps,
-      enabledStats
+      enabledStats,
+      assumeMasterwork
     } = this.props;
     const { rowHeight, rowWidth, rowColumns } = this.state;
 
@@ -141,6 +143,7 @@ export default class GeneratedSets extends React.Component<Props, State> {
             defs={defs}
             statOrder={statOrder}
             enabledStats={enabledStats}
+            assumeMasterwork={assumeMasterwork}
           />
         ) : sets.length > 0 ? (
           <WindowScroller ref={this.windowScroller}>
@@ -166,6 +169,7 @@ export default class GeneratedSets extends React.Component<Props, State> {
                     defs={defs}
                     statOrder={statOrder}
                     enabledStats={enabledStats}
+                    assumeMasterwork={assumeMasterwork}
                   />
                 )}
                 scrollTop={scrollTop}